Refactors experimental code into a directive
Refactors experimental code for drag and drop file import
into the fileDropTarget directive.
