+++ views/index.ejs
@@ -46,7 +46,7 @@
 
 
 </head>
-<body ng-controller="Base">
+<body ng-controller="Base" file-drop-target>
 
   <form method="POST" id="downloader">
     <input type="hidden" name="name">
